feat: Undo・Redoのテストの実装方針を決めて実装する
