+++ docs/細かい設計方針.md
@@ -124,3 +124,8 @@ export type HogeFugaType = z.infer<typeof hogeFugaSchema>;
 
 複数のAI開発ツール間でコーディング規約を統一管理するため、rulesyncを試験的に導入しています。
 `.rulesync/`ディレクトリに統一ルールを配置し、各ツール向けの設定ファイルは自動生成します。
+
+## e2eテストの実装対象について
+
+エディタ上で広範囲で使われている基礎となる機能のテストは、専門のテストファイルは作らず、各機能のテストに含める形で実装します。
+これには例えばVuexの初期化やUndo・Redoが含まれます。
