+++ tests/e2e/browser/テキスト追加・削除・入れ替え.spec.ts
@@ -1,63 +1,85 @@
-import { test, expect, type Locator } from "@playwright/test";
+import { test, expect } from "@playwright/test";
 
 import { gotoHome, navigateToMain } from "../navigators";
 import { fillAudioCell, validateInput } from "./utils";
+import { undo, redo } from "./undo";
+import { ensureNotNullish } from "@/type/utility";
 
 test.beforeEach(gotoHome);
-
-async function getCenter(locator: Locator) {
-  const box = (await locator.boundingBox()) || {
-    x: 0,
-    y: 0,
-    width: 0,
-    height: 0,
-  };
-  return { x: box.x + box.width / 2, y: box.y + box.height / 2 };
-}
 
 test("テキストの追加・入れ替え・削除", async ({ page }) => {
   // デフォルトでaudioCellは一つなのを確認
   await navigateToMain(page);
-  await expect(
-    page.getByRole("button").filter({ hasText: "add" }),
-  ).toBeVisible();
-  expect(await page.locator(".audio-cell").count()).toBe(1);
-  // ３つAudioCellを追加したらAudioCellが４つになるのを確認
-  await page.getByRole("button").filter({ hasText: "add" }).click();
-  await page.getByRole("button").filter({ hasText: "add" }).click();
-  await page.getByRole("button").filter({ hasText: "add" }).click();
-  await page.waitForTimeout(100);
-  await fillAudioCell(page, 0, "foo");
-  await fillAudioCell(page, 1, "bar");
-  await fillAudioCell(page, 2, "baz");
-  expect(await page.locator(".audio-cell").count()).toBe(4);
-  // 一番上のAudioCellを削除しもともと２番めだったものが一番上に来てAudioCellが３つになることを確認
-  await page.locator(".audio-cell").first().hover();
-  await page
-    .getByRole("button")
-    .filter({ hasText: "delete_outline" })
-    .first()
-    .click();
-  await page.waitForTimeout(100);
-  expect(await page.locator(".audio-cell").count()).toBe(3);
-  await validateInput(page.locator(".audio-cell input").first(), "bar");
 
-  // ドラッグして一番上と２番めに上のものを入れ替えて、入れ替わってることを確認
-  const dragFrom = await getCenter(
-    page.locator(".audio-cell .icon-container").first(),
-  );
-  const dragTo = await getCenter(
-    page.locator(".audio-cell .icon-container").nth(1),
-  );
-  await page.mouse.move(dragFrom.x, dragFrom.y);
-  await page.mouse.down();
-  await page.mouse.move(dragTo.x, dragTo.y);
-  await page.mouse.up();
+  await test.step("AudioCellを4つ追加する", async () => {
+    await expect(
+      page.getByRole("button").filter({ hasText: "add" }),
+    ).toBeVisible();
+    expect(await page.locator(".audio-cell").count()).toBe(1);
+    // ３つAudioCellを追加したらAudioCellが４つになるのを確認
+    await page.getByRole("button").filter({ hasText: "add" }).click();
+    await page.getByRole("button").filter({ hasText: "add" }).click();
+    await page.getByRole("button").filter({ hasText: "add" }).click();
+    await page.waitForTimeout(100);
+    await fillAudioCell(page, 0, "foo");
+    await fillAudioCell(page, 1, "bar");
+    await fillAudioCell(page, 2, "baz");
+    expect(await page.locator(".audio-cell").count()).toBe(4);
+  });
 
-  await page.waitForTimeout(100);
-  await validateInput(page.locator(".audio-cell input").first(), "baz");
-  await page.waitForTimeout(100);
-  await validateInput(page.locator(".audio-cell input").nth(1), "bar");
+  await test.step("削除ボタンを押すとAudioCellが削除される", async () => {
+    await page.locator(".audio-cell").first().hover();
+    await page
+      .getByRole("button")
+      .filter({ hasText: "delete_outline" })
+      .first()
+      .click();
+    expect(await page.locator(".audio-cell").count()).toBe(3);
+    await validateInput(page.locator(".audio-cell input").first(), "bar");
+  });
+
+  await test.step("ドラッグするとAudioCellの順番が入れ替わる", async () => {
+    const cells = page.locator(".audio-cell");
+
+    const fromHandle = cells
+      .nth(1)
+      .getByRole("button", { name: /^キャラクターとスタイルを変更、/ });
+    const toHandle = cells
+      .nth(2)
+      .getByRole("button", { name: /^キャラクターとスタイルを変更、/ });
+
+    const fromBox = ensureNotNullish(await fromHandle.boundingBox());
+    const toBox = ensureNotNullish(await toHandle.boundingBox());
+
+    await page.mouse.move(
+      fromBox.x + fromBox.width / 2,
+      fromBox.y + fromBox.height / 2,
+    );
+    await page.mouse.down();
+    await page.mouse.move(
+      toBox.x + toBox.width / 2,
+      toBox.y + toBox.height / 2,
+      { steps: 10 },
+    );
+    await page.mouse.up();
+
+    await validateInput(page.locator(".audio-cell input").first(), "baz");
+    await validateInput(page.locator(".audio-cell input").nth(1), "bar");
+  });
+
+  await test.step("Undoしたら元の値に戻る", async () => {
+    await undo(page);
+
+    await validateInput(page.locator(".audio-cell input").first(), "bar");
+    await validateInput(page.locator(".audio-cell input").nth(1), "baz");
+  });
+
+  await test.step("Redoしたら入れ替え後の値に戻る", async () => {
+    await redo(page);
+
+    await validateInput(page.locator(".audio-cell input").first(), "baz");
+    await validateInput(page.locator(".audio-cell input").nth(1), "bar");
+  });
 });
 
 test("選択中のAudioCellを削除しても正しくフォーカスが移動する", async ({
